Put Cardano types into separate modules
