+++ src/ZkFold/Symbolic/Cardano/Types/Transaction.hs
@@ -1,0 +1,31 @@
+module ZkFold.Symbolic.Cardano.Types.Transaction where
+
+import           Prelude                              hiding (Bool, Eq, length, splitAt, (*), (+))
+
+import           ZkFold.Base.Algebra.Basic.Number
+import           ZkFold.Base.Data.Vector
+import           ZkFold.Symbolic.Cardano.Types.Input  (Input)
+import           ZkFold.Symbolic.Cardano.Types.Output (Output)
+import           ZkFold.Symbolic.Compiler
+import           ZkFold.Symbolic.Data.UTCTime
+
+newtype Transaction inputs rinputs outputs tokens datum a = Transaction
+    ( Vector rinputs (Input tokens datum a)
+    , (Vector inputs (Input tokens datum a)
+    , (Vector outputs (Output tokens datum a)
+    , (UTCTime a, UTCTime a)
+    )))
+
+deriving instance
+    ( Arithmetic a
+    , KnownNat inputs
+    , KnownNat rinputs
+    , KnownNat outputs
+    , KnownNat tokens
+    ) => SymbolicData a (Transaction inputs rinputs outputs tokens datum (ArithmeticCircuit a))
+
+txInputs :: Transaction inputs rinputs outputs tokens datum a -> Vector inputs (Input tokens datum a)
+txInputs (Transaction (_, (is, _))) = is
+
+txOutputs :: Transaction inputs rinputs outputs tokens datum a -> Vector outputs (Output tokens datum a)
+txOutputs (Transaction (_, (_, (os, _)))) = os
