+++ src/ZkFold/Symbolic/Cardano/Types/Value.hs
@@ -1,0 +1,13 @@
+module ZkFold.Symbolic.Cardano.Types.Value where
+
+import           Prelude                            hiding (Bool, Eq, length, splitAt, (*), (+))
+
+import           ZkFold.Base.Algebra.Basic.Number
+import           ZkFold.Base.Data.Vector
+import           ZkFold.Symbolic.Compiler
+import           ZkFold.Symbolic.Data.ByteString
+import           ZkFold.Symbolic.Data.UInt
+
+newtype Value n a = Value (Vector n (ByteString 224 a, (ByteString 256 a, UInt 64 a)))
+
+deriving instance (Arithmetic a, KnownNat n) => SymbolicData a (Value n (ArithmeticCircuit a))
