+++ zkfold-base.cabal
@@ -114,18 +114,24 @@ library
       ZkFold.Base.Protocol.Commitment.KZG
       ZkFold.Base.Protocol.NonInteractiveProof
       ZkFold.Prelude
+      ZkFold.Symbolic.Algorithms.Hash.SHA2
+      ZkFold.Symbolic.Algorithms.Hash.SHA2.Constants
+      ZkFold.Symbolic.Algorithms.Hash.MiMC
+      ZkFold.Symbolic.Algorithms.Hash.MiMC.Constants
       ZkFold.Symbolic.Cardano.Contracts.BatchTransfer
       ZkFold.Symbolic.Cardano.Types
+      ZkFold.Symbolic.Cardano.Types.Address
+      ZkFold.Symbolic.Cardano.Types.Input
+      ZkFold.Symbolic.Cardano.Types.Output
+      ZkFold.Symbolic.Cardano.Types.OutputRef
+      ZkFold.Symbolic.Cardano.Types.Transaction
+      ZkFold.Symbolic.Cardano.Types.Value
       ZkFold.Symbolic.Cardano.UPLC
       ZkFold.Symbolic.Cardano.UPLC.Builtins
       ZkFold.Symbolic.Cardano.UPLC.Inference
       ZkFold.Symbolic.Cardano.UPLC.Inference.Internal
       ZkFold.Symbolic.Cardano.UPLC.Term
       ZkFold.Symbolic.Cardano.UPLC.Type
-      ZkFold.Symbolic.Algorithms.Hash.SHA2
-      ZkFold.Symbolic.Algorithms.Hash.SHA2.Constants
-      ZkFold.Symbolic.Algorithms.Hash.MiMC
-      ZkFold.Symbolic.Algorithms.Hash.MiMC.Constants
       ZkFold.Symbolic.Compiler
       ZkFold.Symbolic.Compiler.ArithmeticCircuit
       ZkFold.Symbolic.Compiler.ArithmeticCircuit.Combinators
@@ -187,6 +193,7 @@ test-suite zkfold-base-test
       Tests.Arithmetization.Test1
       Tests.Arithmetization.Test2
       Tests.Arithmetization.Test3
+      Tests.Arithmetization.Test4
       Tests.Binary
       Tests.ByteString
       Tests.Field
@@ -197,7 +204,6 @@ test-suite zkfold-base-test
       Tests.Pairing
       Tests.Permutations
       Tests.Plonk
-      Tests.Scripts.LockedByTxId
       Tests.SHA2
       Tests.UInt
       Tests.Univariate
